apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "freecad_api"
version = "0.1.0"
description = "FreeCAD tabanlı üretim platformu API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.111.0",
  "uvicorn[standard]==0.30.1",
  "pydantic==2.7.1",
  "pydantic-settings==2.3.3",
  "sqlalchemy==2.0.30",
  "alembic==1.13.2",
  "psycopg2-binary==2.9.9",
  "celery==5.4.0",
  "redis==5.0.4",
  "boto3==1.34.122",
  "minio==7.2.7",
  "opentelemetry-sdk==1.25.0",
  "opentelemetry-exporter-otlp==1.25.0",
  "opentelemetry-instrumentation-fastapi==0.46b0",
  "opentelemetry-instrumentation-requests==0.46b0",
  "python-json-logger==2.0.7",
  "python-multipart==0.0.9",
  "pyjwt[crypto]==2.8.0",
  "passlib[bcrypt]==1.7.4",
  "tenacity==8.3.0",
  "prometheus-client==0.20.0",
  "prometheus-fastapi-instrumentator==6.1.0",
  "mcubes==0.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.2.2",
  "pytest-cov==5.0.0",
  "httpx==0.27.0",
  "ruff==0.4.8",
  "black==24.4.2",
  "isort==5.13.2",
  "bandit==1.7.8",
  "mypy==1.10.0",
  "hypothesis==6.103.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "C90"]
ignore = []

[tool.setuptools]
license-files = []

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["tests*"]


